clkmgr: Proxy merge modules
Merge Proxy 'msgq_tport' and 'message' into the 'Client' module.
The three modules are handling the same things:
Messages, sessions, transmitters and listener.

As they are not really separated,
 just handle them in one module simplify handling.
