+++ clkmgr/proxy/client.cpp
@@ -11,6 +11,9 @@
 
 #include "proxy/client.hpp"
 #include "proxy/connect_ptp4l.hpp"
+#include "proxy/connect_msg.hpp"
+#include "proxy/notification_msg.hpp"
+#include "proxy/subscribe_msg.hpp"
 #ifdef HAVE_LIBCHRONY
 #include "proxy/connect_chrony.hpp"
 #endif
@@ -116,6 +119,64 @@ Transmitter *Client::getTxContext(sessionId_t sessionId)
 Client *Client::getClient(sessionId_t sessionId)
 {
     return sessionMap.count(sessionId) > 0 ? sessionMap[sessionId].get() : nullptr;
+}
+
+bool Client::init()
+{
+    // Register messages we recieve from client side
+    reg_message_type<ProxyConnectMessage, ProxySubscribeMessage>();
+    // ProxyNotificationMessage - Proxy send it only, never send from client
+    Listener &rxContext = Listener::getSingleListenerInstance();
+    PrintDebug("Initializing Proxy listener Queue ...");
+    if(!rxContext.init(mqProxyName, MAX_CLIENT_COUNT)) {
+        PrintError("Initializing Proxy listener queue failed");
+        return false;
+    }
+    PrintDebug("Proxy listener queue opened");
+    return true;
+}
+
+void Client::NotifyClients(size_t timeBaseIndex,
+    vector<sessionId_t> &subscribedClients,
+    vector<sessionId_t> &sessionIdToRemove)
+{
+    ProxyNotificationMessage *pmsg = new ProxyNotificationMessage();
+    if(pmsg == nullptr) {
+        PrintErrorCode("[clkmgr::notify_client] notifyMsg is nullptr !!");
+        return;
+    }
+    // Release message on function ends
+    unique_ptr<ProxyNotificationMessage> notifyMsg(pmsg);
+    PrintDebug("[clkmgr::notify_client] notifyMsg creation is OK !!");
+    // Send data for multiple sessions
+    pmsg->setTimeBaseIndex(timeBaseIndex);
+    for(auto it = subscribedClients.begin(); it != subscribedClients.end();) {
+        const sessionId_t sessionId = *it;
+        PrintDebug("Get client session ID: " + to_string(sessionId));
+        pmsg->set_sessionId(sessionId);
+        if(!pmsg->transmitMessage()) {
+            it = subscribedClients.erase(it);
+            /* Add sessionId into the list to remove */
+            sessionIdToRemove.push_back(sessionId);
+        } else
+            ++it;
+    }
+}
+
+void Client::RemoveClients(const vector<sessionId_t> &sessionIdToRemove)
+{
+    for(const sessionId_t sessionId : sessionIdToRemove) {
+        ConnectPtp4l::remove_ptp4l_subscriber(sessionId);
+        #ifdef HAVE_LIBCHRONY
+        ConnectChrony::remove_chrony_subscriber(sessionId);
+        #endif
+        Client::RemoveClient(sessionId);
+    }
+}
+
+Transmitter *Transmitter::getTransmitterInstance(sessionId_t sessionId)
+{
+    return Client::getTxContext(sessionId);
 }
 
 class ClientRemoveAll : public End
