+++ clkmgr/proxy/client.hpp
@@ -13,11 +13,14 @@
 #define PROXY_CLIENT_HPP
 
 #include "common/msgq_tport.hpp"
+#include "common/message.hpp"
 
 #include <memory>
 #include <map>
 
 __CLKMGR_NAMESPACE_BEGIN
+
+static const size_t MAX_CLIENT_COUNT = 8;
 
 class Client
 {
@@ -25,13 +28,18 @@ class Client
     sessionId_t m_sessionId = InvalidSessionId;
     std::unique_ptr<Transmitter> m_transmitContext;
     static sessionId_t CreateClientSession(const std::string &id);
+    static void RemoveClient(sessionId_t sessionId);
 
   public:
+    static bool init();
     static sessionId_t connect(sessionId_t sessionId, const std::string &id);
     static bool subscribe(size_t timeBaseIndex, sessionId_t sessionId);
-    static void RemoveClient(sessionId_t sessionId);
-    static Transmitter *getTxContext(sessionId_t sessionId);
+    static void RemoveClients(const std::vector<sessionId_t> &sessionIdToRemove);
+    static void NotifyClients(size_t timeBaseIndex,
+        std::vector<sessionId_t> &subscribedClients,
+        std::vector<sessionId_t> &sessionIdToRemove);
     static Client *getClient(sessionId_t sessionId);
+    static Transmitter *getTxContext(sessionId_t sessionId);
     sessionId_t getSessionId() const { return m_sessionId; }
     Transmitter *getTxContext() { return m_transmitContext.get(); }
 };
